# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moustass API",
        "description": "API sécurisée pour l'application Moustass",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://www.moustass.com/api",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/users/recipients": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Liste des destinataires disponibles",
                "operationId": "7fba79fe81591c3b8907b3487a1f7fb0",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs clients"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "Liste des messages reçus",
                "operationId": "29e5814ac26e9f3abac96878fd74bb77",
                "responses": {
                    "200": {
                        "description": "Liste des messages vocaux reçus"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "Envoyer un message vocal",
                "operationId": "941d583a09ed64becb2c2f4753c22501",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "audio",
                                    "receiver_id"
                                ],
                                "properties": {
                                    "audio": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "receiver_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Message envoyé avec succès"
                    },
                    "400": {
                        "description": "Données invalides"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/messages/{id}/read": {
            "put": {
                "tags": [
                    "Messages"
                ],
                "summary": "Marquer un message comme lu",
                "operationId": "4e6fbf082f97015801770c42b0bfb124",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message marqué comme lu"
                    },
                    "404": {
                        "description": "Message introuvable"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Liste des utilisateurs actifs",
                "operationId": "7e2f5799553ec93c9c43e2a58490d447",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs actifs"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "tags": [
                    "Utilisateurs"
                ],
                "summary": "Voir le profil d'un utilisateur",
                "operationId": "910a0c7cf580e1c605648d82a4e57f46",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Profil de l'utilisateur"
                    },
                    "404": {
                        "description": "Utilisateur introuvable"
                    },
                    "401": {
                        "description": "Non authentifié"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}